refactor memory, address and long

+++ no-95/app/src/main/java/forth/ForthInterpreterOperationsLoop.java
@@ -1,7 +1,7 @@
 package forth;
 
 public interface ForthInterpreterOperationsLoop {
-    void pushLoop(int start, int limit);
+    void pushLoop(long start, long limit);
     boolean incrementLoop();
     void popLoop();
 }
