+++ no-95/app/src/main/java/forth/ForthInterpreterOperationsMemory.java
@@ -1,26 +1,29 @@
 package forth;
 
+import forth.memory.Constant;
+import forth.memory.Variable;
+
 public interface ForthInterpreterOperationsMemory {
     boolean hasVariable(String name);
     Variable defineVariable(String name);
-    Variable defineVariable(String name, Integer length);
+    Variable defineVariable(String name, Long length);
     Variable getVariable(String name);
 
     boolean hasConstant(String name);
     Constant defineConstant(String name);
-    Constant defineConstant(String name, Integer length);
+    Constant defineConstant(String name, Long length);
     Constant getConstant(String name);
 
-    Integer getCell(Integer address);
-    void setCell(Integer address, Integer value);
-    Integer cellAllot(Integer length);
+    Long getCell(Long address);
+    void setCell(Long address, Long value);
+    Long cellAllot(Long length);
 
-    char getCharMemory(Integer address);
-    void setCharMemory(Integer address, char value);
-    void setStringMemory(Integer address, String value);
-    String getStringMemory(Integer address, Integer length);
-    Integer charAllot(Integer length);
-    Integer stringAllot(String value);
+    char getCharMemory(Long address);
+    void setCharMemory(Long address, char value);
+    void setStringMemory(Long address, String value);
+    String getStringMemory(Long address, Long length);
+    Long charAllot(Long length);
+    Long stringAllot(String value);
 
 }
