+++ no-95/app/src/main/java/forth/ForthInterpreterOperationsStandard.java
@@ -3,9 +3,9 @@ package forth;
 import java.util.List;
 
 public interface ForthInterpreterOperationsStandard {
-    void push(Integer token);
-    Integer pop();
-    Integer peek();
+    void push(Long token);
+    Long pop();
+    Long peek();
     void executeWord(String word);
     void executePrint(String string);
     void define(String word, List<ForthInterpreter.Instruction> instructions);
